[build] Fix prepareMobileBuild script
iOS build script was providing invalid options.

Make the options more explicit and make validation more strict

+++ buildSrc/prepareMobileBuild.js
@@ -1,44 +1,60 @@
 import fs from "node:fs"
-import { program } from "commander"
+import { Option, program } from "commander"
 import { fileURLToPath } from "node:url"
-import "zx/globals"
 import path from "node:path"
 
+const TAG = "prepareMobileBuild:"
+
 if (process.argv[1] === fileURLToPath(import.meta.url)) {
-	program.usage("make|dist").arguments("<target>").action(prepareMobileBuild).parse(process.argv)
+	program
+		.description("Prepare built web code for mobile app build")
+		.addOption(new Option("--app <app>", "which app to prepare build for").choices(["mail", "calendar"]))
+		.parse(process.argv)
+	await prepareMobileBuild(program.opts())
 }
 
 /**
  * Removes source maps, icons, HTML files which are not needed for mobile apps.
+ * @param opts {object}
+ * @param opts.app {"mail"|"calendar"}
  */
-export async function prepareMobileBuild(buildType, app) {
-	console.log("prepare mobile build for build type", buildType)
-	let prefix
-	if (["dist", "make"].includes(buildType)) {
-		prefix = app === "mail" ? "build/" : "build-calendar-app/"
-	} else {
-		throw new Error("Unknown build type " + buildType)
+export async function prepareMobileBuild({ app }) {
+	if (app !== "mail" && app !== "calendar") {
+		throw new Error("Required option app: " + app)
+	}
+	console.log(TAG, "prepare mobile build for app", app)
+	const prefix = app === "mail" ? "build/" : "build-calendar-app/"
+	if (!fs.existsSync(prefix)) {
+		throw new Error(`Prefix dir does not exist, likely an error!: ${prefix}`)
 	}
 
-	const wasmpath = prefix + "wasm"
-	if (fs.existsSync(wasmpath)) {
-		console.log("unlinking ", wasmpath)
-		fs.rmSync(wasmpath, { force: true, recursive: true })
-	}
-
-	// remove source map
-	for (let file of fs.readdirSync(prefix)) {
-		if (file.endsWith(".js.map")) {
-			console.log("unlinking ", file)
-			fs.unlinkSync(path.join(prefix, file))
-		}
-	}
+	removeWasm(prefix)
+	removeSourceMaps(prefix)
 
 	const indexHtmlPath = prefix + "index.html"
 	if (fs.existsSync(indexHtmlPath)) {
 		fs.unlinkSync(indexHtmlPath)
-		console.log("rm ", indexHtmlPath)
+		console.log(TAG, "rm", indexHtmlPath)
 	} else {
-		console.log("no file at", indexHtmlPath)
+		console.log(TAG, "no file at", indexHtmlPath)
+	}
+}
+
+/** @param prefix {string} */
+function removeWasm(prefix) {
+	const wasmpath = path.join(prefix, "wasm")
+	if (fs.existsSync(wasmpath)) {
+		console.log(TAG, "rm", wasmpath)
+		fs.rmSync(wasmpath, { force: true, recursive: true })
+	}
+}
+
+/** @param prefix {string} */
+function removeSourceMaps(prefix) {
+	for (let file of fs.readdirSync(prefix)) {
+		if (file.endsWith(".js.map")) {
+			console.log(TAG, "rm", file)
+			fs.unlinkSync(path.join(prefix, file))
+		}
 	}
 }
